cinema/cinema-app/requirements.txt
Flask
flask_cors
Jinja2
Flask-RESTful
jsonschema
six
mysql-connector-python